fix: move and refactor commented tests
moved commented tests from scriptspec to pronunciationanalysisSpec. Refactored those tests and deleted empty scriptSpec file.
