+++ .eslintrc
@@ -1,5 +1,7 @@
 root: true
 
+extends: d-centralize/es6
+
 env:
   browser: true
   es6: true
