+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "autobahn": "0.10.1",
     "guid": "0.0.12",
+    "node-fetch": "^1.6.1",
     "pcmjs": "^0.0.2"
   },
   "devDependencies": {
@@ -27,15 +28,22 @@
     "babelify": "^7.3.0",
     "browserify": "^13.1.0",
     "concurrently": "^2.2.0",
-    "eslint": "^3.5.0",
+    "eslint": "^3.2.2",
     "eslint-config-d-centralize": "^0.1.0",
+    "jasmine": "^2.5.1",
     "jasmine-ajax": "^3.2.0",
+    "jasmine-as-promised": "0.0.8",
     "jasmine-core": "^2.4.1",
+    "jasmine-node": "^1.14.5",
+    "jasmine-promises": "^0.4.1",
     "karma": "^1.2.0",
     "karma-browserify": "^5.1.0",
     "karma-cli": "^1.0.1",
+    "karma-coverage": "^1.1.1",
     "karma-jasmine": "^1.0.2",
     "karma-phantomjs-launcher": "^1.0.1",
-    "watchify": "^3.7.0"
+    "q": "^1.4.1",
+    "watchify": "^3.7.0",
+    "xhr-mock": "^1.6.0"
   }
 }
