+++ test/pronunciationAnalysisSpec.js
@@ -19,8 +19,7 @@ require('jasmine-ajax');
 require('jasmine-as-promised')();
 const autobahn = require('autobahn');
 var Promise = require('es6-promise').Promise;
-const its = require('..');
-var Q = require('q');
+const its = require('../');
 
 describe('Pronunciation Analyisis Websocket API interaction test', function() {
   beforeEach(function() {
@@ -149,6 +148,297 @@ describe('Pronunciation Analyisis Websocket API interaction test', function() {
           {trimStart: 0.15, trimEnd: 0.0});
       }).catch(function(error) {
         fail('No error should be thrown: ' + error);
+      });
+  });
+});
+
+describe('PronunciationAnalyses API interaction test', function() {
+  beforeEach(function() {
+    jasmine.Ajax.install();
+
+    // XXX: jasmine-ajax doesn't support asserting FormData yet.
+    // Workaround by attaching a spy while appending to FormData.
+    // https://github.com/pivotal/jasmine-ajax/issues/51
+    spyOn(FormData.prototype, 'append');
+  });
+
+  afterEach(function() {
+    jasmine.Ajax.uninstall();
+  });
+
+
+  it('should get an existing pronunciation analysis', function() {
+    var api = new its.Sdk({
+      authPrincipal: 'principal',
+      authPassword: 'secret'
+    });
+
+    var challenge = new its.PronunciationChallenge('fb', '4', 'test', new Blob);
+
+    var url = 'https://api.itslanguage.nl/organisations/fb' +
+      '/challenges/pronunciation/4/analyses/5';
+    var audioUrl = 'https://api.itslanguage.nl/download/Ysjd7bUGseu8-bsJ';
+    var content = {
+      id: '5',
+      created: '2014-12-31T23:59:59Z',
+      updated: '2014-12-31T23:59:59Z',
+      audioUrl: audioUrl,
+      studentId: '6'
+    };
+    var fakeResponse = {
+      status: 200,
+      contentType: 'application/json',
+      responseText: JSON.stringify(content)
+    };
+    console.log('stubbing url ' + url);
+    jasmine.Ajax.stubRequest(url).andReturn(fakeResponse);
+
+    var output = api.getPronunciationAnalysis(challenge, '5');
+    expect(output).toEqual(jasmine.any(Promise));
+    return output.then(function(result) {
+
+      var request = jasmine.Ajax.requests.mostRecent();
+      expect(request.url).toBe(url);
+      expect(request.method).toBe('GET');
+      var stringDate = '2014-12-31T23:59:59Z';
+      var student = new its.Student('fb', '6');
+      var analysis = new its.PronunciationAnalysis(challenge, student,
+        '5', new Date(stringDate), new Date(stringDate), audioUrl);
+      expect(result).toEqual(analysis);
+
+    }).catch(function(error) {
+
+      fail('No error should be thrown: ' + error);
+
+    });
+  });
+
+  it('should get a list of existing pronunciation analyses', function() {
+    var api = new its.Sdk({
+      authPrincipal: 'principal',
+      authPassword: 'secret'
+    });
+
+    var challenge = new its.PronunciationChallenge('fb', '4', 'test', new Blob);
+    var audioUrl = 'https://api.itslanguage.nl/download/Ysjd7bUGseu8-bsJ';
+    var content = [{
+      id: '5',
+      created: '2014-12-31T23:59:59Z',
+      updated: '2014-12-31T23:59:59Z',
+      audioUrl: audioUrl,
+      studentId: '6'
+    }, {
+      id: '6',
+      created: '2014-12-31T23:59:59Z',
+      updated: '2014-12-31T23:59:59Z',
+      audioUrl: audioUrl,
+      studentId: '24',
+      score: 7.5,
+      words: [
+        [
+          {
+            graphemes: 'b',
+            score: 0.9,
+            verdict: 'good'
+          },
+          {
+            graphemes: 'o',
+            score: 0.4,
+            verdict: 'bad'
+          },
+          {
+            graphemes: 'x',
+            score: 0.5,
+            verdict: 'moderate'
+          }
+        ]
+      ]
+    }];
+
+    var fakeResponse = {
+      status: 200,
+      contentType: 'application/json',
+      responseText: JSON.stringify(content)
+    };
+
+    var url = 'https://api.itslanguage.nl/organisations/fb' +
+      '/challenges/pronunciation/4/analyses';
+    jasmine.Ajax.stubRequest(url).andReturn(fakeResponse);
+
+    var output = api.listPronunciationAnalyses(challenge, false);
+
+    expect(output).toEqual(jasmine.any(Promise));
+
+    return output
+      .then(function(result) {
+
+        var request = jasmine.Ajax.requests.mostRecent();
+        expect(request.url).toBe(url);
+        expect(request.method).toBe('GET');
+        var stringDate = '2014-12-31T23:59:59Z';
+        var student = new its.Student('fb', '6');
+        var analysis = new its.PronunciationAnalysis(challenge, student,
+          '5', new Date(stringDate), new Date(stringDate));
+        analysis.audioUrl = audioUrl;
+        var student2 = new its.Student('fb', '24');
+        var analysis2 = new its.PronunciationAnalysis(challenge, student2,
+          '6', new Date(stringDate), new Date(stringDate));
+        analysis2.audioUrl = audioUrl;
+        analysis2.score = 7.5;
+        var chunk = [
+          new its.WordChunk('b', 0.9, 'good', []),
+          new its.WordChunk('o', 0.4, 'bad', []),
+          new its.WordChunk('x', 0.5, 'moderate', []),
+        ];
+        var word = new its.Word(chunk);
+        analysis2.words = [word];
+        console.log('result is ' + result);
+        expect(result).toEqual(jasmine.any(Array));
+        expect(result.length).toBe(2);
+        expect(result).toEqual([analysis, analysis2]);
+
+      })
+      .catch(function(error) {
+
+        fail('No error should be thrown: ' + error);
+
+      });
+  });
+
+  it('should get a detailed list of pronunciation analyses', function() {
+    var api = new its.Sdk({
+      authPrincipal: 'principal',
+      authPassword: 'secret'
+    });
+
+    var challenge = new its.PronunciationChallenge('fb', '4', 'test', new Blob);
+    var url = 'https://api.itslanguage.nl/organisations/fb' +
+      '/challenges/pronunciation/4/analyses?detailed=true';
+
+    var audioUrl = 'https://api.itslanguage.nl/download/Ysjd7bUGseu8-bsJ';
+    var content = [{
+      id: '5',
+      created: '2014-12-31T23:59:59Z',
+      updated: '2014-12-31T23:59:59Z',
+      audioUrl: audioUrl,
+      studentId: '6'
+    }, {
+      id: '6',
+      created: '2014-12-31T23:59:59Z',
+      updated: '2014-12-31T23:59:59Z',
+      audioUrl: audioUrl,
+      studentId: '24',
+      score: 7.5,
+      words: [
+        [
+          {
+            graphemes: 'b',
+            phonemes: [
+              {
+                ipa: 'b',
+                score: 0.9,
+                verdict: 'good',
+                start: 0.11,
+                end: 0.22
+              }
+            ],
+            score: 0.9,
+            verdict: 'good'
+          },
+          {
+            graphemes: 'o',
+            phonemes: [
+              {
+                ipa: '\u0251',
+                score: 0.4,
+                verdict: 'bad'
+              }
+            ],
+            score: 0.4,
+            verdict: 'bad'
+          },
+          {
+            graphemes: 'x',
+            phonemes: [
+              {
+                ipa: 'k',
+                score: 0.4,
+                verdict: 'bad'
+              },
+              {
+                ipa: 's',
+                score: 0.6,
+                verdict: 'moderate'
+              }
+            ],
+            score: 0.5,
+            verdict: 'moderate'
+          }
+        ]
+      ]
+    }];
+
+    var fakeResponse = {
+      status: 200,
+      contentType: 'application/json',
+      responseText: JSON.stringify(content)
+    };
+
+    jasmine.Ajax.stubRequest(url).andReturn(fakeResponse);
+
+    var output = api.listPronunciationAnalyses(challenge, true);
+    expect(output).toEqual(jasmine.any(Promise));
+
+    return output
+      .then(function(result) {
+
+        var request = jasmine.Ajax.requests.mostRecent();
+        expect(request.url).toBe(url);
+        expect(request.method).toBe('GET');
+        var stringDate = '2014-12-31T23:59:59Z';
+        var student = new its.Student('fb', '6');
+        var analysis = new its.PronunciationAnalysis(challenge, student,
+          '5', new Date(stringDate), new Date(stringDate));
+        analysis.audioUrl = audioUrl;
+        var student2 = new its.Student('fb', '24');
+        var analysis2 = new its.PronunciationAnalysis(challenge, student2,
+          '6', new Date(stringDate), new Date(stringDate));
+        analysis2.audioUrl = audioUrl;
+        analysis2.score = 7.5;
+        var phoneme1 = new its.Phoneme('b', 0.9);
+        phoneme1.verdict = 'good';
+        phoneme1.start = 0.11;
+        phoneme1.end = 0.22;
+        var phonemes1 = [phoneme1];
+        var phoneme2 = new its.Phoneme('\u0251', 0.4);
+        phoneme2.verdict = 'bad';
+        var phonemes2 = [phoneme2];
+        var phoneme4 = new its.Phoneme('k', 0.4);
+        phoneme4.verdict = 'bad';
+        var phoneme5 = new its.Phoneme('s', 0.6);
+        phoneme5.verdict = 'moderate';
+        var phonemes3 = [
+          phoneme4,
+          phoneme5
+        ];
+        var chunk = [
+          new its.WordChunk('b', 0.9, 'good', phonemes1),
+          new its.WordChunk('o', 0.4, 'bad', phonemes2),
+          new its.WordChunk('x', 0.5, 'moderate', phonemes3),
+        ];
+        var word = new its.Word(chunk);
+        analysis2.words = [word];
+        expect(result).toEqual(jasmine.any(Array));
+        expect(result.length).toBe(2);
+        expect(result[0]).toEqual(analysis);
+        expect(result[1]).toEqual(analysis2);
+        expect(result).toEqual([analysis, analysis2]);
+
+      })
+      .catch(function(error) {
+
+        fail('No error should be thrown: ' + error);
+
       });
   });
 });
