+++ test/pronunciationChallengeSpec.js
@@ -19,7 +19,7 @@ require('jasmine-ajax');
 require('jasmine-as-promised')();
 const autobahn = require('autobahn');
 var Promise = require('es6-promise').Promise;
-const its = require('..');
+const its = require('../');
 
 describe('PronunciationChallenge object test', function() {
   it('should require all required fields in constructor', function() {
